+++ spoonacular-kotlin/src/main/java/com/beetlestance/spoonacular_kotlin/models/request/recipe/RequestVisualizeRecipeNutrition.kt
@@ -16,13 +16,13 @@
 package com.beetlestance.spoonacular_kotlin.models.request.recipe
 
 import com.squareup.moshi.Json
-import java.math.BigDecimal
 
 /**
  * @param ingredientList The ingredient list of the recipe, one ingredient per line.
  * @param servings The number of servings.
  * @param defaultCss Whether the default CSS should be added to the response. (optional)
- * @param showBacklink Whether to show a backlink to spoonacular. If set false, this call counts against your quota. (optional)
+ * @param showBacklink Whether to show a backlink to spoonacular. If set false, this call
+ * counts against your quota. (optional)
  * */
 data class RequestVisualizeRecipeNutrition(
 
@@ -30,7 +30,7 @@ data class RequestVisualizeRecipeNutrition(
     val ingredientList: String,
 
     @Json(name = "servings")
-    val servings: BigDecimal,
+    val servings: Int,
 
     @Json(name = "defaultCss")
     val defaultCss: Boolean?,
